Avoid warnings caused by sign

+++ src/ModbusRTUClient.cpp
@@ -35,7 +35,7 @@ ModbusRTUClientClass::~ModbusRTUClientClass()
 {
 }
 
-int ModbusRTUClientClass::begin(unsigned long baudrate, uint16_t config)
+int ModbusRTUClientClass::begin(unsigned long baudrate, int config)
 {
   modbus_t* mb = modbus_new_rtu(baudrate, config);
